Moved httpRequest script to a higher node
The new script location better refects the ui elements that control and
are controlled by the httpRequest.

Also added types to the nodes + nodepaths in a more reasonable way.
Example:

	export(NodePath) var lineEditPath
	onready var lineEdit: LineEdit = get_node(lineEditPath)

This combination seems to be a good pattern to keep the script clean of
precise nodepaths, so that they can be connected via the node inspector,
rather than typed directly in the script.