Converted Game into its own scene, separate from the UI around it
It's easier to operate on the game when it is its own scene,
since the tab UI cluttered up the editor view. Additionally, it
fixed a minor issue that happens with the TileMap editor, where
the tiles didn't match up with the editor's grid guides.  Finally,
the game _should_ be in its own scene, because it is a separation
of concerns.