Added a script to TileMap to experiment with auto-drawing tiles
The script currently makes no changes on the game itself when it runs,
but it is the beginning of a script that will take a 2d array of tile
layout data, and automatically draw them onto the screen when the game
is loaded. This will later be used for loading game maps from the
server.