Moved around Repository structure; added Bow shooting
Pressing the space bar now will shoot a bow in front of the player
with a simple animation. No arrow object is actually created. The
bow can be shot again after the animation finishes.

In this commit, the repo was also reorganized to keep the root
src directory less cluttered. Files are now closer to where they
are used in the scene trees. For example, src/game now holds
scripts and scenes related to the game itself, whereas
src/container holds the files related to the tab interface and
http requester script.